desloppify: typed PostureContext replaces posture-string sentinel + folder rename
Two small-refactors:
- Introduce PostureContext enum in mcp/server.rs with Account(Posture)
  and Infrastructure variants. Replaces the 'infrastructure' string
  sentinel that was passed as a raw String to run_with_audit_envelope
  and emit_tool_start. The enum renders to the same audit-record
  string via as_str() so on-disk audit format is unchanged, but the
  type now prevents silently mis-spelling the sentinel elsewhere.
- Rename crates/rimap-server/src/tools/folder_mgmt.rs to
  folder_management.rs (the only abbreviated filename in tools/).
