+++ crates/rimap-server/src/tools/folder_management.rs
